+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 3.1.2 - 2020-05-26
 
 - Switch to `object-keys` for retrieving the object keys in `|without`
+- Clean up null input of without
 
 ## 3.1.1 - 2020-03-24